Rename: enums 이동
